--- server/data-import/README.md
@@ -1,8 +1,0 @@
-# Import Data
-
-1. Save the desired spreadsheet to import as a .csv file in the `server/data-import` directory
-1. Update lines 6-8 in `import-google-sheets.py` with the name of the .csv file you would like to 
-import and the name of the .json file to create
-1. run `python3 import-google-sheets.py` from the data-import directory
-1. `yarn data` from the root `help-assist-her` directory
-1. `yarn geocode`
