+++ server/util/actions/index.js
@@ -1,13 +1,13 @@
-const createPregnancyCenter = require('./actions/create-pregnancy-center')
-const updateFqhc = require('./actions/update-fqhc')
-const releaseDocuments = require('./actions/release-documents')
+const createPregnancyCenter = require('./create-pregnancy-center')
+const updateFqhc = require('./update-fqhc')
+const releaseDocuments = require('./release-documents')
 const {
 	updateFqhcDoNotList,
 	updateFqhcOutOfBusiness,
 	updatePregnancyCenterDoNotList,
 	updatePregnancyCenterOutOfBusiness,
-} = require('./actions/update-fields')
-const updatePregnancyCenter = require('./actions/update-pregnancy-center')
+} = require('./update-fields')
+const updatePregnancyCenter = require('./update-pregnancy-center')
 
 module.exports = {
 	createPregnancyCenter,
